feat: curve value
