+++ src/main/kotlin/net/ccbluex/liquidbounce/config/types/BindValue.kt
@@ -1,0 +1,32 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2025 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.ccbluex.liquidbounce.config.types
+
+import net.ccbluex.liquidbounce.utils.input.InputBind
+
+class BindValue(
+    name: String,
+    aliases: Array<String> = emptyArray(),
+    defaultValue: InputBind,
+) : Value<InputBind>(name, aliases, defaultValue, ValueType.BIND) {
+    override fun setByString(string: String) {
+        get().bind(string)
+    }
+}
