+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/ModuleJumpEffect.kt
@@ -39,7 +39,7 @@ object ModuleJumpEffect : ClientModule("JumpEffect", Category.RENDER) {
     private val innerColor by color("InnerColor", Color4b(0, 255, 4, 0))
     private val outerColor by color("OuterColor", Color4b(0, 255, 4, 89))
 
-    private val animCurve by curve("AnimCurve", Easing.QUAD_OUT)
+    private val animCurve by easing("AnimCurve", Easing.QUAD_OUT)
 
     private val hueOffsetAnim by int("HueOffsetAnim", 63, -360..360)
