+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/ModuleZoom.kt
@@ -60,7 +60,7 @@ object ModuleZoom : ClientModule("Zoom", Category.RENDER, bindAction = InputBind
         tree(Scroll)
     }
 
-    private val transition by curve("Transition", Easing.QUAD_IN)
+    private val transition by easing("Transition", Easing.QUAD_IN)
     private val durationFactor by float("DurationFactor", 2f, 0f..10f, "x")
 
     private val chronometer = Chronometer()
